work

+++ backend/src/main/java/com/ll/coffee/order/OrderDto.java
@@ -14,9 +14,9 @@ public class OrderDto {
     private String email;
     private String address;
     private int postalCode;
-    private Map<Integer, Integer> orders;
+    private Map<Long, Integer> orders;
 
-    public OrderDto(Order order, Map<Integer, Integer> orders) {
+    public OrderDto(Order order, Map<Long, Integer> orders) {
         this.id = order.getId();
         this.email = order.getEmail();
         this.address = order.getAddress();
